requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
apache-tvm==0.11.1
astor==0.8.1
asttokens==3.0.0
astunparse==1.6.3
attrs==23.2.0
backcall==0.2.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2024.2.2
charset-normalizer==3.3.2
cloudpickle==3.1.1
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.1
dataclasses==0.6
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.2.0
flatbuffers==24.3.25
fonttools==4.51.0
gast==0.3.3
gitdb==4.0.12
GitPython==3.1.44
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.63.0
h5py==2.10.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
ipython==8.12.3
jdcal==1.4.1
jedi==0.19.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
more-itertools==10.2.0
mpmath==1.3.0
mxnet==1.9.1
networkx==3.1
numpy==1.23.4
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.19.2
opencv-python==4.11.0.86
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==24.0
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pluggy==1.5.0
prompt_toolkit==3.0.51
protobuf==3.20.2
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydlcuda @ file:///usr/local/denglin/sdk/python/pydlcuda-1.0.0-cp38-cp38-manylinux2014_x86_64.whl#sha256=4f73812f464a035938d5e177f7bff5ff93761f4630ee7b0e43f600a46474b822
pydlnne @ file:///usr/local/denglin/sdk/python/pydlnne-1.0.0-cp38-cp38-manylinux2014_x86_64.whl#sha256=e7fd8cd361e10075377134ae70147c09eb37a5115863e8bcf1538fa2f876e64e
Pygments==2.19.2
PyMySQL==1.1.0
pyparsing==3.1.2
pytest==8.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.19.3
scipy==1.4.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.12
synr==0.6.0
tensorboard==2.2.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
termcolor==2.4.0
tf-slim==1.1.0
thop==0.1.1.post2209072238
tifffile==2023.7.10
tomli==2.0.1
torch==1.9.1+cpu
torchaudio==0.9.1
torchvision==0.10.1+cpu
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2025.2
ultralytics==8.3.159
ultralytics-thop==2.0.14
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.18.1
